requirements.txt
flask==3.0.0
gunicorn==21.2.0
mysql-connector-python==8.2.0
opencv-python-headless==4.8.1.78
opencv-contrib-python-headless==4.8.1.78
numpy==1.26.2
bcrypt==4.1.2
Pillow==10.1.0
python-dotenv==1.0.0
